SignatureDataConversion_dotNet
Project migration to Visual Studio 2015 / .NET Core
