--- extras/SignatureDataConversion_dotNet/core/.gitignore
@@ -1,1 +1,0 @@
-jayrock/
