+++ extras/SignatureDataConversion_dotNet/src/SignatureDataConversion.Core/project.json
@@ -1,0 +1,17 @@
+{
+  "language": "en",
+  "title": "SignatureDataConversion.Core",
+  "description": "jSignature Tools",
+  "copyright": "Copyright © 2016 by jSignature",
+  "version": "2.0.0-*",
+
+  "dependencies": {
+    "NETStandard.Library": "1.5.0-rc2-24027"
+  },
+
+  "frameworks": {
+    "netstandard1.5": {
+      "imports": "dnxcore50"
+    }
+  }
+}
