+++ extras/SignatureDataConversion_dotNet/test/SignatureDataConversion.Tests/Properties/launchSettings.json
@@ -1,0 +1,7 @@
+{
+  "profiles": {
+    "SignatureDataConversion.Tests": {
+      "commandName": "Project"
+    }
+  }
+}
